Configure TLSv1.3 ciphersuites.
OpenSSL has separate API's for cipher configuration for TLSv1.2 (and below)
and TLSv1.3. Previously, only the former API was called, which left no way
to configure TLSv1.3 ciphersuites. The OpenSSL documentation says that the
configuration strings for the 1.2 and 1.3 API's are compatible (and passing 1.2
ciphers to the 1.3 API will be ignored, and vice versa), so this commit extends
the processing of the `--cipher-list` option to additionally configure TLS 1.3
ciphersuites where available.