Fix potential command injection in docker release workflow
It was possible to run arbitrary commands in the context of the GitHub Actions workflow
by pushing specially crafted tags, e.g. `git tag
'docker/0.0.0-1;$(cat${IFS}/etc/passwd)#'`.
In this example, the workflow would run the `cat` command and print
the contents of `/etc/passwd`.

This requires both knowledge of the workflow code,
the execution environment as well as tag push permission.

Note that git tags are quite restrictive in terms of which characters
are allowed, so one cannot just paste a shell script, and that the tag push
permission would usually imply that an attacker would also be able to
directly modify workflow code.

Nevertheless, as a best practice, we shall try to sanitize any user input that is used in a `run` step.

Further reading:
- https://wearezeta.atlassian.net/wiki/spaces/SC/pages/1216053328/How+to+avoid+Command+Injection+via+Github+Actions+or+CI+jobs+in+general
- https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#example-of-a-script-injection-attack

Related to WPB-9577