Rate-limit 401 Unauthorized responses to prevent abuse/reflection based attacks
Added rate-limiting to 401 Unauthorized responses to prevent abuse of the server
for use in DDoS attacks via traffic reflection and amplification.

This patch works by counting the amount of requests that result in a 401
Unauthorized response and limiting them by IP Address if the occurred in
a specified window of time.

Added new command-line options:

--401-req-limit - Sets the amount of requests that result a 401 response per
                  ratelimt window. If set to 0 disables 401 rate limiting.
--401-window - Sets the size in seconds of the rate-limit window.