Add spdx tags to all source files (#1510)
With notable exceptions of:

src/apps/common/win/*
src/apps/relay/telnet.*

The purpose of this change is to add the SPDX tags from
https://spdx.dev/, which is a linux foundation project, to the source
code.

This provides automated code provenance tools, which are used in setting
up software bill of materials reports, an easy time verifying that the
code license is known and no incompatibilities are present in a
codebase.

No copyright date, author, or license changes are made.

Note also that
https://github.com/coturn/coturn/commit/7e525c8e1c09c0055fb00fbcac091c4d87d19937
is the original commit for the ACME code (acme.h and acme.c) which was
then moved to acme.h and acme.c in this commit
https://github.com/coturn/coturn/commit/d4686750eee289f05e24e9fdcc99c88d4fb217f2
but neither commit indicates what license the ACME code was submitted
as.

https://github.com/coturn/coturn?tab=License-1-ov-file#readme is the
3-clause BSD license, but https://github.com/coturn/coturn/pull/672
documents that the author's intent was for the MIT license. So I've used
the SPDX tag and content of the MIT license for this change.